refactor: Ajuste de nome do dashboard
